Add files via upload
Added in the final changes for the data loads, not just ddl loading.  Takes about an hour, but its worth it.  The data will need to be cleaned up, as it's still the older data load, but the process is now solid up to the AAS tabular model and manual data load of customer data.